requirements.txt
aiogram==3.22.0
python-dotenv==1.0.1
openai>=1.0.0
aiosqlite==0.20.0
prometheus-client==0.20.0